tests/requirements.txt
pytest
pytest-playwright
pytest-asyncio
pytest-lazy-fixtures